requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
Faker==37.1.0
mongoengine==0.29.1
psycopg2-binary==2.9.10
pymongo==3.11.4
pytz==2025.1
sqlparse==0.5.3
tzdata==2025.1
